feat(server/broadcast): broadcast ai command

+++ server/include/angle_direction.h
@@ -8,7 +8,7 @@
 #pragma once
 #include "sys/types.h"
 typedef struct angle_direction_s {
-    int min_angle;
-    int max_angle;
+    double min_angle;
+    double max_angle;
     u_int8_t direction;
 } angle_direction_t;
